pipeline/requirements.txt
requests==2.31.0
simplejson==3.19.1
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
python-dotenv==1.0.0
black==23.9.1